implemented start of clustering

+++ src/Similarity/CosineSimilarity.java
@@ -12,7 +12,7 @@ public class CosineSimilarity extends SimilarityFunction {
         if (main_vector.getMagnitude() == 0 || comp_vector.getMagnitude() == 0) {
             return sum;
         }
-        for (Map.Entry<String, Integer> entry : main_vector.getVector().entrySet()) {
+        for (Map.Entry<String, Double> entry : main_vector.getVector().entrySet()) {
             if (entry.getValue() != 0) {
                 if (comp_vector.getVector().containsKey(entry.getKey()) && comp_vector.getVector().get(entry.getKey()) != 0) {
                     sum += entry.getValue() * comp_vector.getVector().get(entry.getKey());
